change nFile default from 0 to -1
we want to change the default from 0 to -1. We need this so we can deteremine where on disk a blocks dats is. the use of BLOCK_HAVE_DATA determines if we have the data for a given block somewhere on disk and nFile being -1 would indicate data is only in blockdb. nFile >=0 means it is somewhere in a sequential file or possibly sequential file and blockdb.