add cblockindex fields and new disk storage
use additional fields in cblockindex and a new CDbBlockIndex class to store that additional data. we are able to import from old CDiskBlockIndex and this keeps better compatability with the nPoS nFile system that the other bch implementations use. we should be able to swap the client with and ABC node and still be able to startup without any issues.