break if MSG_GRAPHENEBLOCK inv was processed
We break from the loop if we processed an inv for a graphen block.

Edit the comments as to the reason why we break for various inv types.
We don't want to process more than one high priority inv so we don't get
stuck holding cs_main while we process potentially 2000 inv's.