Fix an old request manager bug.
This bug was manifesting easily because of the introduction of the
graphene timer.  The fix is just to put the item back into the
availableFrom list , if we didn't actually request the item. This
ensures we don't lose our sources while we wait for any of the timers
to lapse.