Remove code relating to connect-graphene and connect-graphene-force
This functionality is hopelessly broken in RequestBlock(). While it
could no doubt be made to work again, the amount of work and logic
required far outweighs any benefit. In the past these functions were
used only in the very beginning of XTHIN's history and was primarily
used to test over the GFC. They were superceded by AddNode() and really
no longer have any useful purpose. It is also likely that
connect-thinblock and connect-thinblock-force will also be removed
at a future date.