There is no need to add min mempool bytes to the size of the graphene block.
When sending the graphene block there is a check to see if the graphene
block is larger than the full blocksize. At this point we just do a
straight comparison and don't need to add the 8 bytes of the min mempool
byte size.

Also, use the more efficient GetBlockSize() rather than re-serializing
the full block.