Add MSG_THINBLOCK into INV processing so we can request by inv.
This was erroneously removed earlier along with MSG_XTHINBLOCK. The
MSG_XTHINBLOCK is no longer supported through INV but MSG_THINBLOCK
still is and is needed.

THINBLOCK is one that we can request using a GETDATA. We can do this
because we only request a full thinblock if an Xthin has failed and as
such we can re-use the stored bloom filter. However, Xthins or Graphene blocks
can not be requested in this way.