A variety of nits
Mostly this is about preventing the log entries that the thintimer
was started and then stopped when in fact the timer was disabled
and there was no necessity for any log statements.

Fix an unrelated nit with pushKV() in unlimited.cpp. This was
causing a problem when compiling in debug.

A couple of other log nits