Reject blocks by current size only during reconstruction
Since we only used shared pointers for transactions in the mempool, and only
begin recontruction of the block once we have all the necessary data,
it is no longer needed to be able to track bytes used for
multiple thintype blocks that are still in flight and only partially
reconstructed. This commit simplifies the entire procedure and simply
checks if the block has gone over the byte limit during reconstuction and
if so then cleanup the data and disconnet the peer.