Keep all coins from newly found blocks in the dbcache when configured.
Create a default setting such that when the dbcache is configure to
be more than 2GB, then when the dbcache is flushed to disk we will
save any dirty coins that are not fully spent in memory. This will still
allow for more marginal nodes to save the memory over time but will
make for a better performance for higher end nodes.