Fix a few bugs where we flush to often or not enough
In txadmission.cpp we need to check first if the cache is full before
attempting a trim cycle otherwise we end up trimming after each tx
is entered into the mempool.

Fix bug where nCoinCacheMaxSize is less than nMaxCacheIncreaseSinceLastFlush
which causes an underflow error and subsequently we never trim anything
during our trim cycle.

Remove the 0.995 ratio when calculating when to do a full flush, just
flush when we exceed the nCoinCacheMaxSize. This is something that
was there when we did txadmission in a different way and had to trim
transactions from the cache randomly every time we added one after
the cache was full. This is not needed now.